Performance update for Close() function
Updated the Closer implementation to be more performant. Benchmarks now
run up to twice as fast as before. I expect this is because the Close()
function occupies much more time than any other function, according to
the cpu profiling.

Two changes were made: Instead of checking every iteration whether a
string is being read, we now only check after a buffer refresh or when a
quote is encountered. Also, we now bypass the feedq/feed/next system and
read from the buffer directly, which is less safe but much faster.