Minor numeric parsing improvement
We might use the strconv.ParseFloat function when parsing a number.
However, that function takes a string argument, and the number to parse
is stored in a byte slice. Converting from one to the other requires a
new alloc and a string copy.

Now, we use the unsafe library to cast the byte slice directly to a
string, removing the need for an extra alloc and copy. However, this
causes the buffer to escape to the heap, so a workaround is employed to
keep it on the stack. Because of this, the extra alloc no longer occurs.