Lazy number parsing
Strings, Arrays, and Objects are parsed lazily, as their contents are
needed. On the other hand, Bools, Nulls, and Numbers are parsed
immediately when their JsonValue objects are created. Now, however,
Numbers are parsed lazily when ValueNum() is called. This has two
important effects:

You can no longer read a Number from a structure and ignore it; you must
call ValueNum() on the Number in order to parse its value and gain
access to the next item in the structure, even if you just discard the
value afterward.

Alternatively to calling ValueNum(), you can now instead call Close() on
a Number, which discards it quickly without wasting time on parsing.